requirements.txt
# GEMP Game Analytics Dependencies

# Database
mysql-connector-python>=8.0.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Utilities (already in stdlib for Python 3.8+)
# dataclasses
# typing
